+++ config.js
@@ -3,7 +3,7 @@ export default {
   creator: 'ArcadiaCorps',
   plugins: '60+',
   website: 'https://arcadiacorps.online',
-  version: '1.8.2',
+  version: '1.8.3',
   prefix: '.',
   timezone: 'America/Lima',
   sessionName: 'principal',